+++ src/pages/api/profile/get-user-profile-and-addresses.ts
@@ -18,7 +18,7 @@ import getTracer from "@src/configs/tracer";
 import { ERRORS_DATA, webErrorFactory } from "@src/errors/web-errors";
 import { authMiddleware } from "@src/middlewares/auth-middleware";
 import { sentryMiddleware } from "@src/middlewares/sentry-middleware";
-import { getProfileAndAddressAccessTokens } from "@src/utils/get-profile-address-access-token";
+import { getProfileAndAddressAccessTokens } from "@src/utils/get-profile-and-address-access-tokens";
 
 const handler: Handler = async (request, response) => {
   const webErrors = {
